- delete user progress hook use directly from thai wht - add check for progress done - add tutorial pics - change loaded variable to localstorage - split initTippy function - change done state to small tick - delete mark as done - change dialog button to thai - add change to progress done when delete demo - add delete demo tippy - change setup progress to appropriate value

+++ thai_wht/setup/demo.py
@@ -375,13 +375,24 @@ def delete_demo(pwd):
             ]
         },
         {
-            'status': 'Adding pro config value',
-            'fail_msg': 'Failed : Adding pro config value',
+            'status': 'Adding pre config value',
+            'fail_msg': 'Failed : Adding pre config value',
             'tasks': [
                 {
                     'fn': adding_pre_config_value,
                     'args': '',
-                    'fail_msg': 'Failed : Adding pro config value'
+                    'fail_msg': 'Failed : Adding pre config value'
+                },
+            ]
+        },
+        {
+            'status': 'Clear setup progress',
+            'fail_msg': 'Failed : Clear setup progress',
+            'tasks': [
+                {
+                    'fn': clear_completed_state,
+                    'args': '',
+                    'fail_msg': 'Failed : Clear setup progress'
                 },
             ]
         },
@@ -452,6 +463,7 @@ def delete_transaction(args):
         'doctype'
         )
     thai_wht_doc = listdir(thai_wht_doc_path)
+    dont_delete_doctype = ['Setup Progress', 'Setup Progress Action']
 
     # loop through all doctype
     for doctype in thai_wht_doc:
@@ -466,8 +478,9 @@ def delete_transaction(args):
                 json_data = json.load(doc_json_value)
                 doctype_name = json_data['name']
 
-                # delete doctype
-                delete_for_doctype(doctype_name)
+                if doctype_name not in dont_delete_doctype:
+                    # delete doctype
+                    delete_for_doctype(doctype_name)
 
 
 def delete_for_doctype(doctype):
@@ -521,6 +534,12 @@ def clear_cache(site_name):
 
 def adding_pre_config_value(args):
     add_fixture(only=['wht_records'])
+
+
+def clear_completed_state(args):
+    setup_progress = frappe.get_single('Setup Progress')
+    setup_progress.done = 1
+    setup_progress.save()
 
 
 def update_site_status_delete_demo(args):
